Separate refinements code and many improvements
