+++ indra/preassembler/__init__.py
@@ -7,6 +7,7 @@ import networkx as nx
 from indra.util import fast_deepcopy
 from indra.statements import *
 from indra.statements import stmt_type as indra_stmt_type
+from .refinement import *
 
 logger = logging.getLogger(__name__)
 
@@ -442,10 +443,8 @@ class Preassembler(object):
         # We again iterate over statements
         ts = time.time()
         confirmation_filter = \
-            RefinementConfirmationFilter(stmts_by_hash,
-                                         ontology=self.ontology,
-                                         refinement_fun=self.refinement_fun,
-                                         split_groups=split_groups)
+            RefinementConfirmationFilter(ontology=self.ontology,
+                                         refinement_fun=self.refinement_fun)
         for stmt_hash, possible_refined_hashes in stmts_to_compare.items():
             refinements = confirmation_filter.apply(stmt_hash, possible_refined_hashes)
             maps += [(stmt_hash, ref) for ref in refinements]
@@ -831,302 +830,5 @@ def _flatten_evidence_for_stmt(stmt, collect_from):
     return list(total_evidence)
 
 
-def default_refinement_fun(st1, st2, ontology, entities_refined):
-    return st1.refinement_of(st2, ontology, entities_refined)
-
-
 def default_matches_fun(st):
     return st.matches_key()
-
-
-# TODO: we could make the agent key function parameterizable with the
-# preassembler to allow custom agent mappings to the ontology.
-def get_agent_key(agent):
-    """Return a key for an Agent for use in refinement finding.
-
-    Parameters
-    ----------
-    agent : indra.statements.Agent or None
-         An INDRA Agent whose key should be returned.
-
-    Returns
-    -------
-    tuple or None
-        The key that maps the given agent to the ontology, with special
-        handling for ungrounded and None Agents.
-    """
-    if isinstance(agent, Event):
-        agent = agent.concept
-    if agent is None:
-        agent_key = None
-    else:
-        agent_key = agent.get_grounding()
-        if not agent_key[0]:
-            agent_key = ('NAME', agent.name)
-    return agent_key
-
-
-def get_relevant_keys(agent_key, all_keys_for_role, ontology):
-    """Return relevant agent keys for an agent key for refinement finding.
-
-    Parameters
-    ----------
-    agent_key : tuple or None
-        An agent key of interest.
-    all_keys_for_role : set
-        The set of all agent keys in a given statement corpus with a
-        role matching that of the given agent_key.
-    ontology : indra.ontology.IndraOntology
-        An IndraOntology instance with respect to which relevant other
-        agent keys are found for the purposes of refinement.
-
-    Returns
-    -------
-    set
-        The set of relevant agent keys which this given agent key can
-        possibly refine.
-    """
-    relevant_keys = {None, agent_key}
-    if agent_key is not None:
-        relevant_keys |= set(ontology.get_parents(*agent_key))
-    relevant_keys &= all_keys_for_role
-    return relevant_keys
-
-
-class RefinementFilter:
-    def __init__(self):
-        self.shared_data = {}
-
-    def initialize(self, stmts_by_hash):
-        self.shared_data['stmts_by_hash'] = stmts_by_hash
-
-    def apply(self, stmt_hash, possible_refinements=None,
-              split_groups=None):
-        pass
-
-
-class OntologyRefinementFilter(RefinementFilter):
-    def __init__(self, ontology):
-        self.ontology = ontology
-        self.shared_data = {}
-
-    def initialize(self, stmts_by_hash):
-        self.shared_data['stmts_by_hash'] = stmts_by_hash
-
-        # Build up data structure of statement hashes by
-        # statement type
-        stmts_by_type = collections.defaultdict(set)
-        for stmt_hash, stmt in stmts_by_hash.items():
-            stmts_by_type[indra_stmt_type(stmt)].add(stmt_hash)
-        stmts_by_type = dict(stmts_by_type)
-
-        # Now iterate over each statement type and build up
-        # data structures for quick filtering
-        for stmt_type, stmts_this_type in stmts_by_type.items():
-
-            # Step 1. initialize data structures
-            roles = stmts_by_hash[
-                next(iter(stmts_this_type))]._agent_order
-            # Mapping agent keys to statement hashes
-            agent_key_to_hash = {}
-            # Mapping statement hashes to agent keys
-            hash_to_agent_key = {}
-            # All agent keys for a given agent role
-            all_keys_by_role = {}
-            for role in roles:
-                agent_key_to_hash[role] = collections.defaultdict(set)
-                hash_to_agent_key[role] = collections.defaultdict(set)
-
-            # Step 2. Fill up the initial data structures in preparation
-            # for identifying potential refinements
-            for sh in stmts_this_type:
-                for role in roles:
-                    agent_keys = self._agent_keys_for_stmt_role(
-                        stmts_by_hash[sh], role)
-                    for agent_key in agent_keys:
-                        agent_key_to_hash[role][agent_key].add(sh)
-                        hash_to_agent_key[role][sh].add(agent_key)
-
-            agent_key_to_hash = dict(agent_key_to_hash)
-            hash_to_agent_key = dict(hash_to_agent_key)
-
-            for role in roles:
-                all_keys_by_role[role] = set(agent_key_to_hash[role].keys())
-
-            # Step 3. Make these available as shared data for the apply step
-            self.shared_data[stmt_type] = {
-                'agent_key_to_hash': agent_key_to_hash,
-                'hash_to_agent_key': hash_to_agent_key,
-                'all_keys_by_role': all_keys_by_role,
-            }
-
-    @staticmethod
-    def _agent_keys_for_stmt_role(stmt, role):
-        """Return a set of agent keys for a statement's agent in a role.
-
-        The agent key is an "anchor" to the ontology being used and positons
-        a statement, via its agent in this role against other statements it
-        may be related to.
-        """
-        agents = getattr(stmt, role)
-        # Handle a special case here where a list=like agent
-        # role can be empty, here we will consider anything else
-        # to be a refinement, hence add a None key
-        if isinstance(agents, list) and not agents:
-            agent_keys = {None}
-        # Generally, we take all the agent keys for a single or
-        # list-like agent role.
-        else:
-            agent_keys = {get_agent_key(agent) for agent in
-                          (agents if isinstance(agents, list)
-                           else [agents])}
-        return agent_keys
-
-    def apply(self, stmt, possibly_refines=None):
-        # Corner case: if this is a new statement that wasn't part of the
-        # initialization, it is possible that it has a type that we've not
-        # seen during initialization at all. In this case, we can assume
-        # there are no refinements for it.
-        stmt_type = indra_stmt_type(stmt)
-        if stmt_type not in self.shared_data:
-            return {}
-
-        # Step 1. Recover relevant parts ot the initialized data
-        stmts_by_hash = self.shared_data['stmts_by_hash']
-        hash_to_agent_key = self.shared_data[stmt_type]['hash_to_agent_key']
-        agent_key_to_hash = self.shared_data[stmt_type]['agent_key_to_hash']
-        all_keys_by_role = self.shared_data[stmt_type]['all_keys_by_role']
-
-        # Step 2. We iterate over all statements and find ones that this one
-        # can refine
-        stmt_hash = stmt.get_hash()
-        relevants = possibly_refines
-        # We now iterate over all the agent roles in the given statement
-        # type
-        for role, hash_to_agent_key_for_role in hash_to_agent_key.items():
-            # If we have seen this statement before during initialization then
-            # we can use its precalculated agent keys, otherwise we
-            # calculate new agent keys for it.
-            if stmt_hash in hash_to_agent_key_for_role:
-                agent_keys = hash_to_agent_key_for_role[stmt_hash]
-            else:
-                agent_keys = self._agent_keys_for_stmt_role(stmt, role)
-
-            # We get all the agent keys in all other statements that the
-            # agent in this given role in this statement can refine.
-            for agent_key in agent_keys:
-                relevant_keys = get_relevant_keys(
-                    agent_key,
-                    all_keys_by_role[role],
-                    self.ontology)
-                # We now get the actual statement hashes that these other
-                # potentially refined agent keys appear in in the given role
-                role_relevant_stmt_hashes = set.union(
-                    *[agent_key_to_hash[role][rel]
-                      for rel in relevant_keys]) - {stmt_hash}
-                # In the first iteration, we initialize the set with the
-                # relevant statement hashes
-                if relevants is None:
-                    relevants = role_relevant_stmt_hashes
-                # In subsequent iterations, we take the intersection of
-                # the relevant sets per role
-                else:
-                    relevants &= role_relevant_stmt_hashes
-
-        # These hashes are now the ones that this statement needs
-        # to be compared against. Importantly, the relationship is in
-        # a well-defined direction so we don't need to test both ways.
-        return relevants
-
-
-class RefinementConfirmationFilter(RefinementFilter):
-    def __init__(self, ontology, refinement_fun=None):
-        self.ontology = ontology
-        self.refinement_fun = refinement_fun if refinement_fun else \
-            default_refinement_fun
-        self.shared_data = {}
-
-    def initialize(self, stmts_by_hash):
-        self.shared_data['stmts_by_hash'] = stmts_by_hash
-        pass
-
-    def apply(self, stmt, possibly_refines=None,
-              split_groups=None):
-        stmts_by_hash = self.shared_data['stmts_by_hash']
-        stmt_hash = stmt.get_hash()
-        refinements = set()
-        # We again iterate over statements
-        ts = time.time()
-        # We use the previously constructed set of statements that this one
-        # can possibly refine
-        for possible_refined_hash in possibly_refines:
-            # We handle split groups here to only check refinements between
-            # statements that are in different groups to compare
-            if not split_groups or split_groups.get(stmt_hash) != \
-                    split_groups[possible_refined_hash]:
-                # And then do the actual comparison. Here we use
-                # entities_refined=True which means that we assert that
-                # the entities, in each role, are already confirmed to
-                # be "compatible" for refinement, and therefore, we
-                # don't need to again confirm this (i.e., call "isa") in
-                # the refinement_of function.
-                ref = self.refinement_fun(
-                    stmt,
-                    stmts_by_hash[possible_refined_hash],
-                    ontology=self.ontology,
-                    # NOTE: here we assume that the entities at this point
-                    # are definitely refined due to the use of an
-                    # ontology-based pre-filter. If this is not the case
-                    # for some reason then it is the responsibility of the
-                    # user-supplied self.refinement_fun to disregard the
-                    # entities_refined argument.
-                    entities_refined=True)
-                if ref:
-                    refinements.add(possible_refined_hash)
-        return refinements
-
-
-def ontology_refinement_filter(stmts_by_hash, stmts_to_compare, ontology):
-    """Return possible refinement relationships based on an ontology.
-
-    Parameters
-    ----------
-    stmts_by_hash : dict
-        A dict whose keys are statement hashes that point to the
-        (deduplicated) statement with that hash as a value.
-    stmts_to_compare : dict or None
-        A dict of existing statements to compare that will be further
-        filtered down in this function and then returned.
-    ontology : indra.ontology.IndraOntology
-        An IndraOntology instance iwth respect to which this
-        filter is applied.
-
-    Returns
-    -------
-    dict
-        A dict whose keys are statement hashes and values are sets
-        of statement hashes that can potentially be refined by the
-        statement identified by the key.
-    """
-    logger.info('Finding ontology-based refinements for %d statements'
-                % len(stmts_by_type[stmt_type]))
-    ts = time.time()
-    ont_filter = OntologyRefinementFilter(ontology)
-    ont_filter.initialize(stmts_by_hash)
-
-    stmts_to_compare = {stmt_hash: ont_filter.apply(
-        stmt_hash, possible_refinements=stmts_to_compare.get(stmt_hash))}
-    te = time.time()
-    logger.debug('Identified ontology-based possible refinements in %.2fs'
-                 % (te-ts))
-    # Make an empty dict to make sure we don't return a None
-    if stmts_to_compare is None:
-        stmts_to_compare = {}
-    return stmts_to_compare
-
-
-def bio_ontology_refinement_filter(stmts_by_hash, stmts_to_compare):
-    """An ontology refinement filter that works with the INDRA BioOntology."""
-    from indra.ontology.bio import bio_ontology
-    return ontology_refinement_filter(stmts_by_hash, stmts_to_compare,
-                                      ontology=bio_ontology)
